+++ FishMMO-DB/FishMMO-DB/Entities/ChatEntity.cs
@@ -11,8 +11,8 @@ namespace FishMMO_DB.Entities
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
 		public long ID { get; set; }
 		public long CharacterID { get; set; }
-		public int WorldServerID { get; set; }
-		public int SceneServerID { get; set; }
+		public long WorldServerID { get; set; }
+		public long SceneServerID { get; set; }
 		public DateTime TimeCreated { get; set; }
 		public byte Channel { get; set; }
 		public string Message { get; set; }
